// web/src/api/models/TypesVRRPInstanceStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TypesVRRPPeer } from './TypesVRRPPeer';
import {
    TypesVRRPPeerFromJSON,
    TypesVRRPPeerFromJSONTyped,
    TypesVRRPPeerToJSON,
    TypesVRRPPeerToJSONTyped,
} from './TypesVRRPPeer';

/**
 * 
 * @export
 * @interface TypesVRRPInstanceStatus
 */
export interface TypesVRRPInstanceStatus {
    /**
     * 
     * @type {number}
     * @memberof TypesVRRPInstanceStatus
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof TypesVRRPInstanceStatus
     */
    interface: string;
    /**
     * 
     * @type {string}
     * @memberof TypesVRRPInstanceStatus
     */
    master_ip?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesVRRPInstanceStatus
     */
    name?: string;
    /**
     * 
     * @type {Array<TypesVRRPPeer>}
     * @memberof TypesVRRPInstanceStatus
     */
    peers?: Array<TypesVRRPPeer>;
    /**
     * 
     * @type {number}
     * @memberof TypesVRRPInstanceStatus
     */
    priority: number;
    /**
     * 
     * @type {string}
     * @memberof TypesVRRPInstanceStatus
     */
    state: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof TypesVRRPInstanceStatus
     */
    vips?: Array<string>;
}

/**
 * Check if a given object implements the TypesVRRPInstanceStatus interface.
 */
export function instanceOfTypesVRRPInstanceStatus(value: object): value is TypesVRRPInstanceStatus {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('interface' in value) || value['interface'] === undefined) return false;
    if (!('priority' in value) || value['priority'] === undefined) return false;
    if (!('state' in value) || value['state'] === undefined) return false;
    return true;
}

export function TypesVRRPInstanceStatusFromJSON(json: any): TypesVRRPInstanceStatus {
    return TypesVRRPInstanceStatusFromJSONTyped(json, false);
}

export function TypesVRRPInstanceStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesVRRPInstanceStatus {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'interface': json['interface'],
        'master_ip': json['master_ip'] == null ? undefined : json['master_ip'],
        'name': json['name'] == null ? undefined : json['name'],
        'peers': json['peers'] == null ? undefined : ((json['peers'] as Array<any>).map(TypesVRRPPeerFromJSON)),
        'priority': json['priority'],
        'state': json['state'],
        'vips': json['vips'] == null ? undefined : json['vips'],
    };
}

export function TypesVRRPInstanceStatusToJSON(json: any): TypesVRRPInstanceStatus {
    return TypesVRRPInstanceStatusToJSONTyped(json, false);
}

export function TypesVRRPInstanceStatusToJSONTyped(value?: TypesVRRPInstanceStatus | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'interface': value['interface'],
        'master_ip': value['master_ip'],
        'name': value['name'],
        'peers': value['peers'] == null ? undefined : ((value['peers'] as Array<any>).map(TypesVRRPPeerToJSON)),
        'priority': value['priority'],
        'state': value['state'],
        'vips': value['vips'],
    };
}